pyproject.toml
[tool.poetry]
name = "flantastic_backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^1.1.2"
flask-blueprint = "^1.3.0"
flask-sqlalchemy = "^2.4.4"
flask-httpauth = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
